backend: factor out common structs in model code
prepping to hack on these by hopefully making there be fewer places to fix the same bug

rename
